deepClone issue and whyline
SQUASHED: AUTO-COMMIT-doc-journal-2023-05-04.md-deepClone_works.png,AUTO-COMMIT-doc-journal-2023-05-04.md-index.md,AUTO-COMMIT-src-babylonian-programming-editor-babylonian-programming-editor.js,AUTO-COMMIT-src-babylonian-programming-editor-demos-x.js,AUTO-COMMIT-src-babylonian-programming-editor-tests-ast-test.js,AUTO-COMMIT-src-babylonian-programming-editor-utils-ast.js,AUTO-COMMIT-src-babylonian-programming-editor-utils-defaults.js,AUTO-COMMIT-src-babylonian-programming-editor-worker-ast-worker.js,AUTO-COMMIT-src-client-reactive-reactive-jsx-babel-plugin-jsx-lively.js,AUTO-COMMIT-src-client-search-roots.js,AUTO-COMMIT-src-components-demo-lively-whyline.js,AUTO-COMMIT-src-components-demo-lively-whyline.png,AUTO-COMMIT-src-components-tools-lively-ast-inspector.png,

+++ src/client/search-roots.js
@@ -27,11 +27,13 @@ export default class SearchRoots {
     return FileIndex.current().addDirectory(url)     
   }
 
-  static removeSearchRoot(url) {
+  static removeSearchRoot(url, cleanIndex=true) {
     var roots = this.getSearchRoots();
     roots = roots.filter(ea => ea != url)
     this.setSearchRoots(roots)
-    FileIndex.current().removeDirectory(url)     
+    if (cleanIndex) {
+      FileIndex.current().removeDirectory(url)     
+    }
   }
 
   static isSearchRoot(url) {
